Merge front e back

+++ backend-EasyPark/src/main/java/com/backend/EasyPark/entities/Fabricante.java
@@ -17,11 +17,9 @@ public class Fabricante {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String modelo;
     private String marca;
     private int ano;
-
-
 
 }
